feat: rename project from ecoCode to creedengo
