+++ RuleTests/Creedengo/GCI82.VariableCanBeMadeConstant.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class VariableCanBeMadeConstant
 {
