+++ RuleTests/Creedengo/GCI83.ReplaceEnumToStringWithNameOf.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class ReplaceEnumToStringWithNameOf
 {
