+++ RuleTests/Creedengo/GCI86.GCCollectShouldNotBeCalled.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class GCCollectShouldNotBeCalled
 {
