+++ RuleTests/Creedengo/GCI87.UseListIndexer.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class UseListIndexer
 {
