+++ RuleTests/Creedengo/GCI88.DisposeResourceAsynchronusly.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 [SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope", Justification = "Valid warning but we don't need it here")]
 internal static class DisposeResourceAsynchronusly
