+++ RuleTests/Creedengo/GCI91.UseWhereBeforeOrderBy.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class UseWhereBeforeOrderBy
 {
