+++ RuleTests/Creedengo/GCI92.UseLengthToTestEmptyStrings.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class UseLengthToTestEmpyString
 {
