+++ RuleTests/Creedengo/GCI93.ReturnTaskDirectly.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleTests.EcoCode;
+﻿namespace RuleTests.Creedengo;
 
 internal static class ReturnTaskDirectly
 {
