+++ creedengo-csharp-test-project.csproj
@@ -3,8 +3,8 @@
 	<PropertyGroup>
 		<Authors>Green Code Initiative</Authors>
 		<Company>Green Code Initiative</Company>
-		<Description>A project that contains code with warnings that should be raised by the EcoCode analyzers.</Description>
-		<Copyright>Copyright EcoCode © 2024</Copyright>
+		<Description>A project that contains code with warnings that should be raised by the Creedengo analyzers.</Description>
+		<Copyright>Copyright Creedengo © 2024</Copyright>
 
 		<TargetFramework>net8.0</TargetFramework>
 		<LangVersion>latest</LangVersion>
@@ -15,7 +15,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="EcoCode" Version="1.5.1" />
+	  <PackageReference Include="Creedengo" Version="1.5.1" />
 	</ItemGroup>
 
 </Project>
